Adding entities

+++ data/src/main/java/dk/nodes/template/models/dto/Price.kt
@@ -1,6 +1,6 @@
 package dk.nodes.template.models.dto
 
 data class Price(
-    val price: Double = 0.0,
+    val price: Float = 0f,
     val type: String = ""
 )
